+++ aladhan/endpoints.py
@@ -28,13 +28,13 @@ SPECIAL_DAYS = BASE + "specialDays"  # todo
 ISLAMIC_MONTHS = BASE + "islamicMonths"  # todo
 
 # Date Converters
-H_TO_G = BASE + "hToG"  # todo
-G_TO_H = BASE + "gToH"  # todo
-G_TO_H_CALENDAR = BASE + "gToHCalendar/%d/%d?adjustment=%d"  # todo
-H_TO_G_CALENDAR = BASE + "hToGCalendar/%d/%d?adjustment=%d"  # todo
+H_TO_G = BASE + "hToG"
+G_TO_H = BASE + "gToH"
+G_TO_H_CALENDAR = BASE + "gToHCalendar/%d/%d?adjustment=%d"
+H_TO_G_CALENDAR = BASE + "hToGCalendar/%d/%d?adjustment=%d"
 ISLAMIC_YEAR_FROM_G_FOR_RAMADAN = (
-    BASE + "islamicYearFromGregorianForRamadan"
-)  # todo
+    BASE + "islamicYearFromGregorianForRamadan/%d"
+)
 
 # Holidays
 NEXT_HIJRI_HOLIDAY = BASE + "nextHijriHoliday"  # todo